Eigen2 fix -- add casts to floats.
git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@1863 af78ad9f-9711-0410-ba57-a8eff4914ec4

+++ libavogadro/src/engines/iso.cpp
@@ -629,10 +629,10 @@ namespace Avogadro
         m_stepSize = 0.10;
       }
     }
-    m_min = m_grid->cube()->min();
+    m_min = m_grid->cube()->min().cast<float>();
 
     // Work out the max coordinate too
-    m_max = m_grid->cube()->max();
+    m_max = m_grid->cube()->max().cast<float>();
 
     // Right now we are just using one tessellation method
     m_tessellation=&IsoGen::vMarchCube1;
